requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-dotenv>=0.19.0
firebase-admin>=6.2.0
google-api-python-client>=2.0.0
langchain>=0.0.350
requests>=2.31.0
langchain-community
faiss-cpu
openai
google-auth-httplib2
google-auth-oauthlib
pytest>=7.0.0
pytest-asyncio>=0.21.0 
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.0.0